task: created separate FieldTypeNumber component
